Refactor #62: extract match/progress helpers to shared module
Move _build_match_fn, _build_progress_fn, and _format_skip_breakdown
from cli/commands/import_cmd.py to cli/_match_helpers.py so the upcoming
`bookery add` command can reuse them. No behavior change.
