+++ src/bookery/cli/commands/import_cmd.py
@@ -6,13 +6,17 @@ from pathlib import Path
 import click
 from rich.console import Console
 
+from bookery.cli._match_helpers import (
+    build_match_fn,
+    build_progress_fn,
+    format_skip_breakdown,
+)
 from bookery.cli.options import db_option
 from bookery.core.config import get_library_root
 from bookery.core.dedup import filter_redundant_mobis
-from bookery.core.importer import ImportResult, MatchFn, MatchResult, ProgressFn, import_books
+from bookery.core.importer import MatchFn, import_books
 from bookery.db.catalog import LibraryCatalog
 from bookery.db.connection import DEFAULT_DB_PATH, open_library
-from bookery.metadata.types import BookMetadata
 
 console = Console()  # TODO: move Console() inside command for testability
 
@@ -83,116 +87,6 @@ def _convert_mobis(
     )
     return epub_files
 
-
-def _build_match_fn(
-    output_dir: Path, quiet: bool, threshold: float,
-) -> MatchFn:
-    """Build a match callback that runs the full metadata pipeline.
-
-    Imports match-pipeline dependencies lazily so the import command
-    doesn't pay for them when --match is not used.
-    """
-    from bookery.cli.review import ReviewSession
-    from bookery.core.pipeline import match_one
-    from bookery.metadata.http import BookeryHttpClient
-    from bookery.metadata.openlibrary import OpenLibraryProvider
-
-    http_client = BookeryHttpClient()
-    provider = OpenLibraryProvider(http_client=http_client)
-    review = ReviewSession(
-        console=console,
-        quiet=quiet,
-        threshold=threshold,
-        lookup_fn=provider.lookup_by_url,
-    )
-
-    def match_fn(
-        extracted: BookMetadata, epub_path: Path,
-    ) -> MatchResult | None:
-        result = match_one(epub_path, provider, review, output_dir)
-
-        if not quiet and result.normalization and result.normalization.was_modified:
-            console.print(
-                f"  [dim]Normalized:[/dim] {result.normalization.normalized.title}"
-            )
-
-        if result.status == "matched" and result.metadata is not None:
-            if not quiet:
-                console.print(
-                    f"  [green]Written:[/green] {result.output_path}"
-                )
-            return MatchResult(
-                metadata=result.metadata, output_path=result.output_path,
-            )
-
-        if result.status == "error" and not quiet:
-            console.print(
-                f"  [red]Write failed:[/red] {result.error}"
-            )
-
-        return None
-
-    return match_fn
-
-
-def _format_skip_breakdown(result: ImportResult) -> str:
-    """Format a skip count with breakdown by reason."""
-    if result.skipped == 0:
-        return ""
-
-    parts = []
-    if result.skipped_hash:
-        parts.append(f"{result.skipped_hash} hash")
-    if result.skipped_metadata:
-        # Break metadata skips down by specific reason
-        reason_counts: dict[str, int] = {}
-        for detail in result.skip_details:
-            if detail.reason in ("isbn", "title_author"):
-                label = detail.reason.replace("_", "+")
-                reason_counts[label] = reason_counts.get(label, 0) + 1
-        parts.extend(f"{count} {reason}" for reason, count in reason_counts.items())
-
-    breakdown = f" ({', '.join(parts)})" if parts else ""
-    return f"{result.skipped} skipped{breakdown}"
-
-
-def _build_progress_fn() -> ProgressFn:
-    """Build a per-file progress callback for Rich console output."""
-
-    def on_progress(
-        path: Path,
-        title: str,
-        author: str,
-        status: str,
-        reason: str | None,
-        existing_id: int | None,
-    ) -> None:
-        label = f"{title} — {author}" if title and author else path.name
-        if status == "added":
-            console.print(f"  [green]✓[/green] {label}")
-        elif status == "skipped" and reason:
-            reason_label = reason.replace("_", "+")
-            id_suffix = f", #{existing_id}" if existing_id else ""
-            console.print(
-                f"  [yellow]⊘[/yellow] {label} — "
-                f"[dim]skipped (duplicate: {reason_label}{id_suffix})[/dim]"
-            )
-        elif status == "forced" and reason:
-            reason_label = reason.replace("_", "+")
-            id_suffix = f", #{existing_id}" if existing_id else ""
-            console.print(
-                f"  [yellow]⚠[/yellow] {label} — "
-                f"[dim]imported (duplicate: {reason_label}{id_suffix})[/dim]"
-            )
-        elif status == "error":
-            console.print(f"  [red]✗[/red] {path.name} — [red]{reason}[/red]")
-        elif status == "move_failed":
-            console.print(
-                f"  [yellow]⚠[/yellow] {path.name} — "
-                f"[dim]cataloged but source not removed: {reason}[/dim]"
-            )
-
-    return on_progress
 
 
 
@@ -293,13 +187,14 @@ def import_command(
     library_root = output_dir or get_library_root()
     match_fn: MatchFn | None = None
     if do_match:
-        match_fn = _build_match_fn(
+        match_fn = build_match_fn(
+            console=console,
             output_dir=library_root,
             quiet=quiet,
             threshold=threshold,
         )
 
-    on_progress = _build_progress_fn()
+    on_progress = build_progress_fn(console)
 
     result = import_books(
         epub_files, catalog,
@@ -316,7 +211,7 @@ def import_command(
     if result.added:
         parts.append(f"[green]{result.added} added[/green]")
     if result.skipped:
-        parts.append(f"[yellow]{_format_skip_breakdown(result)}[/yellow]")
+        parts.append(f"[yellow]{format_skip_breakdown(result)}[/yellow]")
     if result.forced:
         parts.append(f"[yellow]{result.forced} forced[/yellow]")
     if result.errors:
